requirements.txt
PyYAML~=6.0
pandas~=1.5.3
openai~=0.27.2
requests~=2.28.2
azure-cognitiveservices-speech~=1.26.0
discord.py~=2.2.2
fasttext-langdetect~=1.0.5
langidentification~=1.2
PyNaCl~=1.5.0